django-vue/requirements.txt
django
django-crispy-forms
django-brutebuster2
djangorestframework
django-cors-headers
